Move to analytics api
